@@ -22,8 +22,8 @@ jint JNI_OnLoad(JavaVM *vm, void *reserved) {
     return JNI_VERSION_1_6;
 }
 
-jstring
-Java_net_idik_lib_cipher_so_CipherCore_getString(JNIEnv *env, jobject instance, jstring key_) {
+extern "C" jstring
+Java_net_idik_lib_cipher_so_CipherCore_getString(JNIEnv *env, jclass instance, jstring key_) {
     const char *key = env->GetStringUTFChars(key_, 0);
     string keyStr(key);
     string value = _map[keyStr];
@@ -31,6 +31,6 @@ Java_net_idik_lib_cipher_so_CipherCore_getString(JNIEnv *env, jobject instance, 
     return env->NewStringUTF(value.c_str());
 }
 
-void Java_net_idik_lib_cipher_so_CipherCore_init(JNIEnv *env, jclass type) {
+extern "C" void Java_net_idik_lib_cipher_so_CipherCore_init(JNIEnv *env, jclass type) {
     LOAD_MAP(_map);
 }
